* code style, warning fix

+++ MSTImagePickerController/MSTImagePicker/Views/MSTPhotoPreviewImageCell.m
@@ -18,16 +18,14 @@
 }
 
 @property (strong, nonatomic) UIScrollView *myScrollView;
-
 @property (strong, nonatomic) UIImageView *imageView;
 @property (strong, nonatomic) PHLivePhotoView *livePhotoView;
-
 @property (strong, nonatomic) UIImageView *liveBadgeImageView;
-
 @end
 
 @implementation MSTPhotoPreviewImageCell
 #pragma mark - Inintialization Method
+
 - (instancetype)initWithFrame:(CGRect)frame {
     if (self = [super initWithFrame:frame]) {
         [self mp_setupSubview];
@@ -36,6 +34,7 @@
 }
 
 #pragma mark - Lazy Load
+
 - (UIScrollView *)myScrollView {
     if (!_myScrollView) {
         self.myScrollView = [[UIScrollView alloc] initWithFrame:CGRectMake(10, 0, self.mst_width - 20, self.mst_height)];
@@ -50,7 +49,7 @@
         _myScrollView.delaysContentTouches = NO;
 
         _myScrollView.delegate = self;
-        
+
         [self addSubview:_myScrollView];
     }
     return _myScrollView;
@@ -60,7 +59,7 @@
     if (!_imageView) {
         self.imageView = [UIImageView new];
         _imageView.clipsToBounds = YES;
-        
+
         [self.myScrollView addSubview:_imageView];
     }
     return _imageView;
@@ -70,7 +69,7 @@
     if (!_livePhotoView) {
         self.livePhotoView = [PHLivePhotoView new];
         _livePhotoView.clipsToBounds = YES;
-        
+
         [self.myScrollView addSubview:_livePhotoView];
     }
     return _livePhotoView;
@@ -85,54 +84,85 @@
         _liveBadgeImageView.translatesAutoresizingMaskIntoConstraints = NO;
         _liveBadgeImageView.hidden = YES;
         [self.livePhotoView addSubview:_liveBadgeImageView];
-        
-        NSLayoutConstraint *leading = [NSLayoutConstraint constraintWithItem:_liveBadgeImageView attribute:NSLayoutAttributeLeading relatedBy:NSLayoutRelationEqual toItem:self.livePhotoView attribute:NSLayoutAttributeLeading multiplier:1 constant:0];
-        NSLayoutConstraint *top = [NSLayoutConstraint constraintWithItem:_liveBadgeImageView attribute:NSLayoutAttributeTop relatedBy:NSLayoutRelationEqual toItem:self.livePhotoView attribute:NSLayoutAttributeTop multiplier:1 constant:0];
-        NSLayoutConstraint *width = [NSLayoutConstraint constraintWithItem:_liveBadgeImageView attribute:NSLayoutAttributeTrailing relatedBy:NSLayoutRelationEqual toItem:self.livePhotoView attribute:NSLayoutAttributeLeading multiplier:1 constant:30];
-        NSLayoutConstraint *height = [NSLayoutConstraint constraintWithItem:_liveBadgeImageView attribute:NSLayoutAttributeBottom relatedBy:NSLayoutRelationEqual toItem:self.livePhotoView attribute:NSLayoutAttributeTop multiplier:1 constant:30];
+
+        NSLayoutConstraint *leading = [NSLayoutConstraint constraintWithItem:_liveBadgeImageView
+                                                                   attribute:NSLayoutAttributeLeading
+                                                                   relatedBy:NSLayoutRelationEqual
+                                                                      toItem:self.livePhotoView
+                                                                   attribute:NSLayoutAttributeLeading
+                                                                  multiplier:1
+                                                                    constant:0];
+        NSLayoutConstraint *top = [NSLayoutConstraint constraintWithItem:_liveBadgeImageView
+                                                               attribute:NSLayoutAttributeTop
+                                                               relatedBy:NSLayoutRelationEqual
+                                                                  toItem:self.livePhotoView
+                                                               attribute:NSLayoutAttributeTop
+                                                              multiplier:1
+                                                                constant:0];
+        NSLayoutConstraint *width = [NSLayoutConstraint constraintWithItem:_liveBadgeImageView
+                                                                 attribute:NSLayoutAttributeTrailing
+                                                                 relatedBy:NSLayoutRelationEqual
+                                                                    toItem:self.livePhotoView
+                                                                 attribute:NSLayoutAttributeLeading
+                                                                multiplier:1
+                                                                  constant:30];
+        NSLayoutConstraint *height = [NSLayoutConstraint constraintWithItem:_liveBadgeImageView
+                                                                  attribute:NSLayoutAttributeBottom
+                                                                  relatedBy:NSLayoutRelationEqual
+                                                                     toItem:self.livePhotoView
+                                                                  attribute:NSLayoutAttributeTop
+                                                                 multiplier:1
+                                                                   constant:30];
         [self.livePhotoView addConstraints:@[leading, top, width, height]];
     }
     return _liveBadgeImageView;
 }
 
 #pragma mark - Setter
+
 - (void)setModel:(MSTAssetModel *)model {
     _model = model;
-    
+
     self.imageView.image = nil;
     self.livePhotoView.livePhoto = nil;
     self.liveBadgeImageView.hidden = YES;
-    
+
     if (model.type == MSTAssetModelMediaTypeLivePhoto && [UIDevice currentDevice].systemVersion.floatValue >= 9.1 && [MSTPhotoConfiguration defaultConfiguration].isCallBackLivePhoto) {
         _isLivePhoto = YES;
-        
-        [[MSTPhotoManager defaultManager] getLivePhotoFromPHAsset:model.asset completionBlock:^(PHLivePhoto *livePhoto, BOOL isDegraded) {
-            if (!isDegraded) {
-                self.livePhotoView.livePhoto = livePhoto;
-                if ([MSTPhotoConfiguration defaultConfiguration].isShowLivePhotoIcon)
-                    self.liveBadgeImageView.hidden = NO;
-                [self mp_resizeSubviews];
-            }
-        }];
+
+        [[MSTPhotoManager defaultManager] getLivePhotoFromPHAsset:model.asset
+                                                  completionBlock:^(PHLivePhoto *livePhoto, BOOL isDegraded) {
+                                                      if (!isDegraded) {
+                                                          self.livePhotoView.livePhoto = livePhoto;
+                                                          if ([MSTPhotoConfiguration defaultConfiguration].isShowLivePhotoIcon)
+                                                              self.liveBadgeImageView.hidden = NO;
+                                                          [self mp_resizeSubviews];
+                                                      }
+                                                  }];
     } else {
         _isLivePhoto = NO;
-        
-        [[MSTPhotoManager defaultManager] getPreviewImageFromPHAsset:model.asset isHighQuality:NO completionBlock:^(UIImage *result, NSDictionary *info, BOOL isDegraded) {
-            self.imageView.image = result;
-            [self mp_resizeSubviews];
-        }];
+
+        [[MSTPhotoManager defaultManager] getPreviewImageFromPHAsset:model.asset
+                                                       isHighQuality:NO
+                                                     completionBlock:^(UIImage *result, NSDictionary *info, BOOL isDegraded) {
+                                                         self.imageView.image = result;
+                                                         [self mp_resizeSubviews];
+                                                     }];
     }
 }
 
 #pragma mark - Instance Methods
+
 - (void)didDisplayed {
     if (!_isLivePhoto) {
-        [[MSTPhotoManager defaultManager] getPreviewImageFromPHAsset:_model.asset isHighQuality:YES completionBlock:^(UIImage *result, NSDictionary *info, BOOL isDegraded) {
-            if (!isDegraded) {
-                self.imageView.image = result;
-                [self mp_resizeSubviews];
-            }
-        }];
+        [[MSTPhotoManager defaultManager] getPreviewImageFromPHAsset:_model.asset
+                                                       isHighQuality:YES
+                                                     completionBlock:^(UIImage *result, NSDictionary *info, BOOL isDegraded) {
+                                                         if (!isDegraded) {
+                                                             self.imageView.image = result;
+                                                             [self mp_resizeSubviews];
+                                                         }
+                                                     }];
     }
 }
 
@@ -155,13 +185,15 @@
         default:
             break;
     }
-    
-    UITapGestureRecognizer *singleTap = [[UITapGestureRecognizer alloc] initWithTarget:self action:@selector(mp_singleTap:)];
+
+    UITapGestureRecognizer *singleTap = [[UITapGestureRecognizer alloc] initWithTarget:self
+                                                                                action:@selector(mp_singleTap:)];
     [self addGestureRecognizer:singleTap];
-    
-    UITapGestureRecognizer *doubleTap = [[UITapGestureRecognizer alloc] initWithTarget:self action:@selector(mp_doubleTap:)];
+
+    UITapGestureRecognizer *doubleTap = [[UITapGestureRecognizer alloc] initWithTarget:self
+                                                                                action:@selector(mp_doubleTap:)];
     doubleTap.numberOfTapsRequired = 2;
-    
+
     [singleTap requireGestureRecognizerToFail:doubleTap];
     [self addGestureRecognizer:doubleTap];
 }
@@ -170,44 +202,44 @@
     if (!_isLivePhoto) {
         self.imageView.mst_origin = CGPointZero;
         self.imageView.mst_width = self.myScrollView.mst_width;
-        
+
         UIImage *image = self.imageView.image;
         if (image.size.height / image.size.width > self.mst_height / self.myScrollView.mst_width) {
-            _imageView.mst_height = floor(image.size.height / (image.size.width / self.myScrollView.mst_width));
+            _imageView.mst_height = (CGFloat)floor(image.size.height / (image.size.width / self.myScrollView.mst_width));
         } else {
             CGFloat height = image.size.height / image.size.width * self.myScrollView.mst_width;
             if (height < 1 || isnan(height)) height = self.mst_height;
-            height = floor(height);
+            height = (CGFloat)floor(height);
             _imageView.mst_height = height;
             _imageView.mst_centerY = self.mst_height / 2;
         }
-        
+
         if (_imageView.mst_height > self.mst_height && _imageView.mst_height - self.mst_height <= 1) {
             _imageView.mst_height = self.mst_height;
         }
-        
+
         _myScrollView.contentSize = CGSizeMake(_myScrollView.mst_width, MAX(_imageView.mst_height, self.mst_height));
         [_myScrollView scrollRectToVisible:self.bounds animated:NO];
         _myScrollView.alwaysBounceVertical = _imageView.mst_height <= self.mst_height ? NO : YES;
     } else {
         self.livePhotoView.mst_origin = CGPointZero;
         self.livePhotoView.mst_width = self.myScrollView.mst_width;
-        
+
         PHLivePhoto *photo = self.livePhotoView.livePhoto;
         if (photo.size.height / photo.size.width > self.mst_height / self.myScrollView.mst_width) {
-            _livePhotoView.mst_height = floor(photo.size.height / (photo.size.width / self.myScrollView.mst_width));
+            _livePhotoView.mst_height = (CGFloat)floor(photo.size.height / (photo.size.width / self.myScrollView.mst_width));
         } else {
             CGFloat height = photo.size.height / photo.size.width * self.myScrollView.mst_width;
             if (height < 1 || isnan(height)) height = self.mst_height;
-            height = floor(height);
+            height = (CGFloat)floor(height);
             _livePhotoView.mst_height = height;
             _livePhotoView.mst_centerY = self.mst_height / 2;
         }
-        
+
         if (_livePhotoView.mst_height > self.mst_height && _livePhotoView.mst_height - self.mst_height <= 1) {
             _livePhotoView.mst_height = self.mst_height;
         }
-        
+
         _myScrollView.contentSize = CGSizeMake(_myScrollView.mst_width, MAX(_livePhotoView.mst_height, self.mst_height));
         [_myScrollView scrollRectToVisible:self.bounds animated:NO];
         _myScrollView.alwaysBounceHorizontal = _livePhotoView.mst_height <= self.mst_height ? NO : YES;
@@ -228,22 +260,24 @@
         CGFloat newZoomScale = _myScrollView.maximumZoomScale;
         CGFloat xSize = self.frame.size.width / newZoomScale;
         CGFloat ySize = self.frame.size.height / newZoomScale;
-        [_myScrollView zoomToRect:CGRectMake(touchPoint.x - xSize/2, touchPoint.y - ySize/2, xSize, ySize) animated:YES];
+        [_myScrollView zoomToRect:CGRectMake(touchPoint.x - xSize / 2, touchPoint.y - ySize / 2, xSize, ySize)
+                         animated:YES];
     }
 }
 
 #pragma mark - UIScrollViewDelegate
+
 // 让UIImageView在UIScrollView缩放后居中显示
 - (void)scrollViewDidZoom:(UIScrollView *)scrollView {
-    CGFloat offsetX = (scrollView.bounds.size.width > scrollView.contentSize.width)?
-    (scrollView.bounds.size.width - scrollView.contentSize.width) * 0.5 : 0.0;
-    CGFloat offsetY = (scrollView.bounds.size.height > scrollView.contentSize.height)?
-    (scrollView.bounds.size.height - scrollView.contentSize.height) * 0.5 : 0.0;
-    
+    CGFloat offsetX = (CGFloat)((scrollView.bounds.size.width > scrollView.contentSize.width) ?
+                (scrollView.bounds.size.width - scrollView.contentSize.width) * 0.5 : 0.0);
+    CGFloat offsetY = (CGFloat)((scrollView.bounds.size.height > scrollView.contentSize.height) ?
+                (scrollView.bounds.size.height - scrollView.contentSize.height) * 0.5 : 0.0);
+
     if (!_isLivePhoto) {
-        self.imageView.center = CGPointMake(scrollView.contentSize.width * 0.5 + offsetX, scrollView.contentSize.height * 0.5 + offsetY);
+        self.imageView.center = CGPointMake((CGFloat)(scrollView.contentSize.width * 0.5 + offsetX), (CGFloat)(scrollView.contentSize.height * 0.5 + offsetY));
     } else {
-        self.livePhotoView.center = CGPointMake(scrollView.contentSize.width * 0.5 + offsetX, scrollView.contentSize.height * 0.5 + offsetY);
+        self.livePhotoView.center = CGPointMake((CGFloat)(scrollView.contentSize.width * 0.5 + offsetX), (CGFloat)(scrollView.contentSize.height * 0.5 + offsetY));
     }
 }
